refactor(auth): move repository packages for better organization
