+++ features/auth/domain/src/commonMain/kotlin/com/parksupark/soomjae/features/auth/domain/repositories/AuthRepository.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.features.auth.domain
+package com.parksupark.soomjae.features.auth.domain.repositories
 
 import arrow.core.Either
 import com.parksupark.soomjae.core.domain.failures.DataFailure
